requirements.txt
# langchain
# langchain_experimental
# tabulate

matplotlib==3.8.0
numpy==1.26.1
openai==0.28.1
pandas==2.1.1
python-dotenv==1.0.0
seaborn==0.13.0
streamlit==1.27.2
